requirements.txt
botocore==1.12.188
boto3==1.9.188
requests==2.18.4
Pillow==6.1.0
apscheduler==3.6.1
beautifulsoup4==4.8.0
tweepy==3.8.0
